% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_use.R
\name{general_use}
\alias{general_use}
\title{General Use}
\usage{
general_use(which, key, tour = "pga", file_format = "json")
}
\arguments{
\item{which}{a character string indicating the type of data to retrieve. Possible options are "player_list", "schedules", and "updates".}

\item{key}{a character string representing the API key required to access the Datagolf API.}

\item{tour}{a character string indicating the golf tour for which to retrieve data. This parameter is only used when which is set to "schedules" or "updates". Default is "pga".}

\item{file_format}{a character string indicating the format in which to retrieve data. This parameter is only used when which is set to "player_list", "schedules", or "updates". Default is "json".}
}
\value{
A data frame containing the requested data.
}
\description{
This function retrieves data from the Datagolf API based on the specified input parameters.
}
