% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_betting_odds.R
\name{historical_betting_odds}
\alias{historical_betting_odds}
\title{Historical Betting Odds}
\usage{
historical_betting_odds(
  which,
  key,
  tour = "pga",
  site = "draftkings",
  event_id = NULL,
  year = "2023",
  file_format = "json",
  market = "win",
  book = "betmgm",
  odds_format = "decimal"
)
}
\arguments{
\item{which}{A character string indicating which type of data to retrieve. Possible values are "ids", "outrights", and "matchups".}

\item{key}{A character string specifying the API key to use.}

\item{tour}{A character string indicating the golf tour to retrieve data for. Defaults to "pga".}

\item{site}{A character string indicating the betting site to retrieve data from. Defaults to "draftkings".}

\item{event_id}{An optional character string indicating the ID of the golf event to retrieve data for.}

\item{year}{A character string indicating the year of the golf event to retrieve data for. Defaults to "2023".}

\item{file_format}{A character string indicating the file format of the API response. Defaults to "json".}

\item{market}{A character string indicating the type of betting market to retrieve data for. Defaults to "win".}

\item{book}{A character string indicating the betting book to retrieve data from. Defaults to "betmgm".}

\item{odds_format}{A character string indicating the format of the betting odds. Defaults to "decimal".}
}
\value{
A data frame containing the historical betting odds for the specified golf events.Get historical outright betting odds for a specific golf event

event_id <- "1234"
outright_odds <- historical_betting_odds("outrights", event_id = event_id)Get historical matchup betting odds for a specific golf event

matchup_odds <- historical_betting_odds("matchups", event_id = event_id)
}
\description{
This function retrieves historical betting odds for golf events from the Datagolf API.
}
