% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_raw_data.R
\name{historical_raw_data}
\alias{historical_raw_data}
\title{Historical Raw Data}
\usage{
historical_raw_data(
  which,
  key,
  tour = "pga",
  site = "draftkings",
  event_id = NULL,
  year = "2023",
  file_format = "json"
)
}
\arguments{
\item{which}{A character string indicating which type of data to retrieve. Possible values are "ids" to retrieve a list of event IDs or "scoring" to retrieve scoring data for a specific tournament event. Default is "ids".}

\item{key}{A character string containing the API key for authentication purposes.}

\item{tour}{A character string indicating the tour to retrieve data from. Default is "pga".}

\item{site}{A character string indicating the data provider site. Default is "draftkings".}

\item{event_id}{A character string indicating the ID of the tournament event to retrieve data for. Default is "403".}

\item{year}{A character string indicating the year of the tournament event to retrieve data for. Default is "2023".}

\item{file_format}{A character string indicating the file format to retrieve the data in. Default is "json".}
}
\value{
A data frame containing the retrieved historical raw data.
}
\description{
This function retrieves historical raw data from the DataGolf API by sending GET requests to specific API endpoints. It can retrieve a list of event IDs or scoring data for a specific tournament event.
}
\examples{
\dontrun{
historical_raw_data(which = "ids", key = "your_api_key")
historical_raw_data(which = "scoring", key = "your_api_key", tour = "lpga", site = "pga", event_id = "1234", year = "2022")
}
}
