% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_predictions.R
\name{model_predictions}
\alias{model_predictions}
\title{Retrieve data from the Data Golf API}
\usage{
model_predictions(
  which,
  key,
  file_format = "json",
  tour = "pga",
  odds_format = "percent",
  event_id = NULL,
  year = "2023",
  display = "value",
  period = "l24",
  site = "draftkings",
  slate = "main",
  dead_heat = "no",
  round = "event_avg"
)
}
\arguments{
\item{which}{A character string indicating which Data Golf API endpoint to query. Must be one of 'dg_rankings', 'pre_predictions', 'pre_predictions_archive', 'skill_decomp', 'skill_ratings', 'detailed_approach_skill', 'fantasy_projection_defaults', 'live_model_predictions', 'live_tournament_stats', or 'live_hole_scoring'.}

\item{key}{A character string indicating the API key to use for authentication.}

\item{file_format}{A character string indicating the format in which to return the data. Must be one of 'json' (default), 'xml', 'csv', 'yaml', or 'text'.}

\item{tour}{A character string indicating the golf tour for which to retrieve data. Only applicable for certain endpoints. Default value is 'pga'.}

\item{odds_format}{A character string indicating the format in which to return the betting odds. Only applicable for certain endpoints. Must be one of 'percent' (default) or 'decimal'.}

\item{event_id}{A character string indicating the ID of the golf event for which to retrieve data. Only applicable for certain endpoints.}

\item{year}{A character string indicating the year of the golf event for which to retrieve data. Only applicable for certain endpoints. Default value is '2023'.}

\item{display}{A character string indicating how to display the skill ratings. Only applicable for certain endpoints. Must be one of 'value' (default) or 'rank'.}

\item{period}{A character string indicating the period over which to calculate the detailed approach skill ratings. Only applicable for certain endpoints. Must be one of 'l4', 'l8', 'l12', or 'l24' (default).}

\item{site}{A character string indicating the daily fantasy sports site for which to retrieve data. Only applicable for certain endpoints. Default value is 'draftkings'.}

\item{slate}{A character string indicating the slate for which to retrieve data. Only applicable for certain endpoints. Default value is 'main'.}

\item{dead_heat}{A character string indicating whether to account for dead heat scenarios in the live model predictions. Only applicable for certain endpoints. Must be one of 'yes' or 'no' (default).}

\item{round}{A character string indicating the round for which to retrieve data. Only applicable for certain endpoints. Must be one of 'event_avg' (default), '1', '2', '3', or '4'.}
}
\value{
A data frame containing the requested data.
}
\description{
This function allows the user to retrieve data from the Data Golf API by specifying the API endpoint to query, and any necessary query parameters.
}
